Create new FirebaseMessagingService to handle push receipts.
