+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableNotificationBuilder.java
@@ -21,6 +21,7 @@ import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLConnection;
+import java.util.Map;
 
 /**
  * Created by David Truong dt@iterable.com
@@ -305,6 +306,17 @@ public class IterableNotificationBuilder extends NotificationCompat.Builder {
 
         return isGhostPush;
     }
+
+//    static boolean isGhostPush(Map<String, String> messageData) {
+//        boolean isGhostPush = false;
+//        if (messageData.containsKey(IterableConstants.ITERABLE_DATA_KEY)) {
+//            String iterableData = messageData.get(IterableConstants.ITERABLE_DATA_KEY);
+//            IterableNotificationData data = new IterableNotificationData(iterableData);
+//            isGhostPush = data.getIsGhostPush();
+//        }
+//
+//        return isGhostPush;
+//    }
 
     /**
      * Returns if the given notification has an empty body
