+++ iterableapi/src/main/java/com/iterable/iterableapi/IterablePushReceiver.java
@@ -28,59 +28,61 @@ public class IterablePushReceiver extends BroadcastReceiver{
     public void onReceive(Context context, Intent intent) {
         String intentAction = intent.getAction();
 
-        if (intentAction.equals(ACTION_FCM_RECEIVE_INTENT)) {
-            IterableLogger.d(TAG, "FCM intent received" + intent);
-            handlePushReceived(context, intent);
-        } else if (intentAction.equals(ACTION_GCM_RECEIVE_INTENT)) {
-            IterableLogger.d(TAG, "GCM intent received" + intent);
-            handlePushReceived(context, intent);
-        }
+//        if (intentAction.equals(ACTION_FCM_RECEIVE_INTENT)) {
+//            IterableLogger.d(TAG, "FCM intent received" + intent);
+//            handlePushReceived(context, intent);
+//        } else if (intentAction.equals(ACTION_GCM_RECEIVE_INTENT)) {
+//            IterableLogger.d(TAG, "GCM intent received" + intent);
+//            handlePushReceived(context, intent);
+//        }
     }
 
-    /**
-     * Handles receiving an incoming push notification from the intent.
-     * @param context
-     * @param intent
-     */
-    private void handlePushReceived(Context context, Intent intent) {
-        if (intent.hasExtra(IterableConstants.ITERABLE_DATA_KEY)) {
-            Bundle extras =  intent.getExtras();
-            if (!IterableNotificationBuilder.isGhostPush(extras)) {
-                if (!IterableNotificationBuilder.isEmptyBody(extras)) {
-                    IterableLogger.d(TAG, "Iterable push received " + extras);
-                    Context appContext = context.getApplicationContext();
-                    PackageManager packageManager = appContext.getPackageManager();
-                    Intent packageIntent = packageManager.getLaunchIntentForPackage(appContext.getPackageName());
-                    ComponentName componentPackageName = packageIntent.getComponent();
-                    String mainClassName = componentPackageName.getClassName();
-                    Class mainClass = null;
-                    try {
-                        mainClass = Class.forName(mainClassName);
-                    } catch (ClassNotFoundException e) {
-                        IterableLogger.w(TAG, e.toString());
-                    }
-
-                    IterableNotificationBuilder notificationBuilder = IterableNotificationBuilder.createNotification(
-                            appContext, intent.getExtras(), mainClass);
-                    new IterableNotificationManager().execute(notificationBuilder);
-                } else {
-                    IterableLogger.d(TAG, "Iterable OS notification push received");
-                }
-            } else {
-                IterableLogger.d(TAG, "Iterable ghost silent push received");
-            }
-        }
-    }
+//    /**
+//     * Handles receiving an incoming push notification from the intent.
+//     * @param context
+//     * @param intent
+//     */
+//    private void handlePushReceived(Context context, Intent intent) {
+//        if (intent.hasExtra(IterableConstants.ITERABLE_DATA_KEY)) {
+//            Bundle extras =  intent.getExtras();
+//            if (!IterableNotificationBuilder.isGhostPush(extras)) {
+//                if (!IterableNotificationBuilder.isEmptyBody(extras)) {
+//                    IterableLogger.d(TAG, "Iterable push received " + extras);
+//                    Context appContext = context.getApplicationContext();
+//                    PackageManager packageManager = appContext.getPackageManager();
+//                    Intent packageIntent = packageManager.getLaunchIntentForPackage(appContext.getPackageName());
+//                    ComponentName componentPackageName = packageIntent.getComponent();
+//                    String mainClassName = componentPackageName.getClassName();
+//                    Class mainClass = null;
+//                    try {
+//                        mainClass = Class.forName(mainClassName);
+//                    } catch (ClassNotFoundException e) {
+//                        IterableLogger.w(TAG, e.toString());
+//                    }
+//
+//                    N
+//
+//                    IterableNotificationBuilder notificationBuilder = IterableNotificationBuilder.createNotification(
+//                            appContext, intent.getExtras(), mainClass);
+//                    new IterableNotificationManager().execute(notificationBuilder);
+//                } else {
+//                    IterableLogger.d(TAG, "Iterable OS notification push received");
+//                }
+//            } else {
+//                IterableLogger.d(TAG, "Iterable ghost silent push received");
+//            }
+//        }
+//    }
 }
 
-class IterableNotificationManager extends AsyncTask<IterableNotificationBuilder, Void, Void> {
-
-    @Override
-    protected Void doInBackground(IterableNotificationBuilder... params) {
-        if ( params != null && params[0] != null) {
-            IterableNotificationBuilder notificationBuilder = params[0];
-            IterableNotificationBuilder.postNotificationOnDevice(notificationBuilder.mContext, notificationBuilder);
-        }
-        return null;
-    }
-}
+//class IterableNotificationManager extends AsyncTask<IterableNotificationBuilder, Void, Void> {
+//
+//    @Override
+//    protected Void doInBackground(IterableNotificationBuilder... params) {
+//        if ( params != null && params[0] != null) {
+//            IterableNotificationBuilder notificationBuilder = params[0];
+//            IterableNotificationBuilder.postNotificationOnDevice(notificationBuilder.mContext, notificationBuilder);
+//        }
+//        return null;
+//    }
+//}
